+++ packages/ui/src/components/playback/PlaybackControls.tsx
@@ -32,6 +32,7 @@ export function PlaybackControls() {
         }
         switch (event.key) {
           case ' ':
+            event.preventDefault();
             player.togglePlayback();
             break;
           case 'ArrowLeft':
@@ -111,7 +112,6 @@ export function PlaybackControls() {
         render={(framerate, paused) => (
           <Input
             title="Current framerate"
-            size={4}
             readOnly
             value={paused ? 'PAUSED' : `${framerate} FPS`}
           />